backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
python-docx==1.1.0
supabase==2.3.4
anthropic==0.18.1
openai==1.12.0
google-cloud-aiplatform==1.43.0
groq==0.4.2
mistralai==0.0.12
numpy==1.26.4
nltk==3.8.1
markdown==3.5.2
tiktoken==0.6.0
langchain==0.1.9
langchain-openai==0.0.8
langchain-anthropic==0.0.6
langchain-google-genai==0.0.9
langchain-mistralai==0.0.5
langchain-groq==0.0.1
chromadb==0.4.22 